refactor: extract refactor autofix commit helper
